aigp_doctor/requirements.txt
torch
torchaudio
torchvision
transformers
gradio
Pillow